+++ docs/CHANGES.md
@@ -10,6 +10,7 @@ dev
  * dhunter: added `--clean-db` option to remove dead entries from project DB
  * dhunter: now yields proper warning when source dir is a symlink
  * dscan: now supports `--force` to overwrite existing database file
+ * dscan: added `--relative-paths` option
 
 v1.2.0 (2019-03-11)
 -------------------